hsi-10879 [BE] Modificar endpoint para obtención de recetas con múltiples comerciales

+++ back-end/hospital-public-api/src/main/java/ar/lamansys/sgh/publicapi/prescription/domain/PrescriptionDosageBo.java
@@ -21,7 +21,7 @@ public class PrescriptionDosageBo {
 
 	private String presentation;
 
-	private Integer presentationQuantity;
+	private Short presentationQuantity;
 
 	private Double quantity;
