+++ back-end/hospital-public-api/src/main/java/ar/lamansys/sgh/publicapi/prescription/infrastructure/input/rest/dto/PrescriptionDosageDto.java
@@ -15,7 +15,7 @@ public class PrescriptionDosageDto {
 
 	private String presentation;
 
-	private Integer presentationQuantity;
+	private Short presentationQuantity;
 
 	private Double quantity;
